Reduce process compat coupling with focused helpers
The process compatibility surface had grown into a mixed env, lifecycle, runtime detection, and command execution module. Keep the existing public process.ts import path as a thin barrel while moving implementation into focused helper modules so future maintenance can happen in smaller lanes.

Constraint: Preserve all existing process.ts exports and cross-runtime behavior for Deno, Node, and Bun callers.\nRejected: Updating downstream imports to helper modules | would broaden the public API change and exceed the split task.\nConfidence: high\nScope-risk: narrow\nTested: deno fmt --check src/platform/compat/process.ts src/platform/compat/process/*.ts; DENO_NO_PACKAGE_JSON=1 deno lint src/platform/compat/process.ts src/platform/compat/process/*.ts; deno check src/platform/compat/process.ts src/platform/compat/process.test.ts src/platform/compat/index.ts src/platform/index.ts; deno test --no-check --allow-all src/platform/compat/process.test.ts src/platform/compat/index.test.ts src/platform/compat/cross-runtime-simple.test.ts; deno task typecheck; lsp_diagnostics on all modified files\nNot-tested: Full deno task test / binary build due task-local refactor scope
